Splitting specific SecuredFieldsProviderHandlers tests out to own file
